MVC04

+++ Company.G03.PL/Controllers/Department.cs
@@ -85,10 +85,22 @@ namespace Company.G03.PL.Controllers
             return View(dept);
         }
         [HttpPost]
-        public IActionResult Delete(Department department)
+        public IActionResult Delete(Department department, [FromRoute] int id)
         {
-            _DepartmentRepository.Delete(department);
-            return RedirectToAction(nameof(Index));
+            if (id !=department.Id)
+            {
+                return BadRequest();
+            }
+            try
+            {
+                _DepartmentRepository.Delete(department);
+                return RedirectToAction(nameof(Index));
+            }
+            catch(System.Exception ex)
+            {
+                ModelState.AddModelError(string.Empty, ex.Message);
+                return View(department);
+            }
         }
     }
 }
